Refactor test code to run in parallel

+++ language/tools/move-cli/tests/cli_tests.rs
@@ -13,6 +13,7 @@ use move_command_line_common::movey_constants::MOVEY_URL;
 use std::fs::File;
 use std::{env, fs, io::Write};
 
+use std::path::PathBuf;
 #[cfg(unix)]
 use std::os::unix::fs::PermissionsExt;
 use std::{path::PathBuf, process::Stdio};
@@ -48,20 +49,18 @@ fn run_metatest() {
     assert!(test::run_all(&path_metatest, &path_cli_binary, true, false).is_ok());
 }
 
-const PACKAGE_PATH: &str = "./tests/upload_tests/valid_package";
-
 #[test]
 fn cross_process_locking_git_deps() {
     let cli_exe = env!("CARGO_BIN_EXE_move");
     let handle = std::thread::spawn(move || {
-        std::process::Command::new(cli_exe)
+        Command::new(cli_exe)
             .current_dir("./tests/cross_process_tests/Package1")
             .args(["package", "build"])
             .output()
             .expect("Package1 failed");
     });
     let cli_exe = env!("CARGO_BIN_EXE_move").to_string();
-    std::process::Command::new(cli_exe)
+    Command::new(cli_exe)
         .current_dir("./tests/cross_process_tests/Package2")
         .args(["package", "build"])
         .output()
@@ -69,9 +68,12 @@ fn cross_process_locking_git_deps() {
     handle.join().unwrap();
 }
 
+const UPLOAD_PACKAGE_PATH: &str = "./tests/upload_tests";
+
 #[test]
 fn upload_package_to_movey_works() {
-    let (home, credential_file) = setup_move_home();
+    let test_path = String::from("/upload_package_to_movey_works");
+    let (home, credential_file) = setup_move_home(&test_path);
     let _ = fs::remove_file(&credential_file);
 
     fs::create_dir_all(&home).unwrap();
@@ -79,62 +81,68 @@ fn upload_package_to_movey_works() {
     let credential_content =
         String::from("[registry]\ntoken=\"eb8xZkyr78FNL528j7q39zcdS6mxjBXt\"\n");
     file.write(&credential_content.as_bytes()).unwrap();
-
-    init_git(PACKAGE_PATH, 0);
+    let package_path = format!("{}/valid_package", UPLOAD_PACKAGE_PATH);
+    init_git(&package_path, 0);
     let cli_exe = env!("CARGO_BIN_EXE_move");
     let output = Command::new(cli_exe)
-        .current_dir(PACKAGE_PATH)
-        .args(["package", "upload", "--test"])
+        .current_dir(&package_path)
+        .args(["package", "upload", "--test", "--test-path", &test_path])
         .output()
         .unwrap();
     assert!(output.status.success());
-    let res_path = format!("{}{}", PACKAGE_PATH, "/request-body.txt");
+    let res_path = format!("{}/request-body.txt", &package_path);
     let data = fs::read_to_string(&res_path).unwrap();
     assert!(data.contains("rev"));
     assert!(data.contains("\"github_repo_url\":\"https://github.com/diem/move\""));
     fs::remove_file(&res_path).unwrap();
 
-    clean_up(&home);
+    clean_up(&home, &package_path);
 }
 
 #[test]
 fn upload_package_to_movey_with_no_remote_should_panic() {
-    init_git(PACKAGE_PATH, 1);
+    let package_path = format!("{}/no_git_remote_package", UPLOAD_PACKAGE_PATH);
+    init_git(&package_path, 1);
     let cli_exe = env!("CARGO_BIN_EXE_move");
     let output = Command::new(cli_exe)
-        .current_dir(PACKAGE_PATH)
+        .current_dir(&package_path)
         .args(["package", "upload", "--test"])
         .output()
         .unwrap();
     assert!(!output.status.success());
     let error = String::from_utf8_lossy(output.stderr.as_slice()).to_string();
     assert!(error.contains("invalid git repository"));
+    clean_up("", &package_path);
 }
 
 #[test]
 fn upload_package_to_movey_with_no_head_commit_id_should_panic() {
-    init_git(PACKAGE_PATH, 2);
+    let package_path = format!("{}/no_commit_id_package", UPLOAD_PACKAGE_PATH);
+    init_git(&package_path, 2);
     let cli_exe = env!("CARGO_BIN_EXE_move");
     let output = Command::new(cli_exe)
-        .current_dir(PACKAGE_PATH)
+        .current_dir(&package_path)
         .args(["package", "upload", "--test"])
         .output()
         .unwrap();
     assert!(!output.status.success());
     let error = String::from_utf8_lossy(output.stderr.as_slice()).to_string();
-    assert!(error.contains("invalid HEAD commit id"));
+    assert!(error.contains("invalid HEAD commit id"), "{}", error);
+    clean_up("", &package_path);
 }
 
 #[test]
 fn upload_package_to_movey_with_no_credential_should_panic() {
-    let (home, credential_file) = setup_move_home();
+    let test_path = String::from("/upload_package_to_movey_with_no_credential_should_panic");
+    let (home, credential_file) = setup_move_home(&test_path);
     let _ = fs::remove_file(&credential_file);
 
-    init_git(PACKAGE_PATH, 0);
+    let package_path = format!("{}/no_credential_package", UPLOAD_PACKAGE_PATH);
+    init_git(&package_path, 0);
     let cli_exe = env!("CARGO_BIN_EXE_move");
     let output = Command::new(cli_exe)
-        .current_dir(PACKAGE_PATH)
-        .args(["package", "upload", "--test"])
+        .current_dir(&package_path)
+        .args(["package", "upload", "--test", "--test-path", &test_path])
         .output()
         .unwrap();
     assert!(!output.status.success());
@@ -144,12 +152,13 @@ fn upload_package_to_movey_with_no_credential_should_panic() {
         Please run `move login` and follow the instructions."
     ));
 
-    clean_up(&home);
+    clean_up(&home, &package_path);
 }
 
 #[test]
 fn upload_package_to_movey_with_bad_credential_should_panic() {
-    let (home, credential_file) = setup_move_home();
+    let test_path = String::from("/upload_package_to_movey_with_bad_credential_should_panic");
+    let (home, credential_file) = setup_move_home(&test_path);
     let _ = fs::remove_file(&credential_file);
 
     fs::create_dir_all(&home).unwrap();
@@ -157,11 +166,12 @@ fn upload_package_to_movey_with_bad_credential_should_panic() {
     let bad_content = String::from("[registry]\ntoken=\n");
     file.write(&bad_content.as_bytes()).unwrap();
 
-    init_git(PACKAGE_PATH, 0);
+    let package_path = format!("{}/bad_credential_package", UPLOAD_PACKAGE_PATH);
+    init_git(&package_path, 0);
     let cli_exe = env!("CARGO_BIN_EXE_move");
     let output = Command::new(cli_exe)
-        .current_dir(PACKAGE_PATH)
-        .args(["package", "upload", "--test"])
+        .current_dir(&package_path)
+        .args(["package", "upload", "--test", "--test-path", &test_path])
         .output()
         .unwrap();
     assert!(!output.status.success());
@@ -171,17 +181,13 @@ fn upload_package_to_movey_with_bad_credential_should_panic() {
         Please run `move login` and follow the instructions."
     ));
 
-    clean_up(&home);
+    clean_up(&home, &package_path);
 }
 
 // flag == 0: all git command are run
 // flag == 1: missing git add remote command
 // flag == 2: missing git commit command
 fn init_git(package_path: &str, flag: i32) {
-    let git_path = format!("{}{}", package_path, "/.git");
-    if Path::new::<String>(&git_path).exists() {
-        fs::remove_dir_all(&git_path).unwrap();
-    }
     Command::new("git")
         .current_dir(package_path)
         .args(&["init"])
@@ -196,12 +202,12 @@ fn init_git(package_path: &str, flag: i32) {
     }
     Command::new("touch")
         .current_dir(package_path)
-        .args(&["simplefile"])
+        .args(&["simple_file"])
         .output()
         .unwrap();
     Command::new("git")
         .current_dir(package_path)
-        .args(&["add", "simplefile"])
+        .args(&["add", "simple_file"])
         .output()
         .unwrap();
     if flag != 2 {
@@ -324,12 +330,19 @@ fn clean_up(move_home: &str) {
     let _ = fs::remove_dir_all(move_home);
 }
 
-fn clean_up(move_home: &str) {
-    let _ = fs::remove_dir_all(move_home);
+fn clean_up(move_home: &str, package_path: &str) {
+    if !move_home.is_empty() {
+        let _ = fs::remove_dir_all(move_home);
+    }
+    fs::remove_file(format!("{}/simple_file", package_path)).unwrap();
+    fs::remove_dir_all(format!("{}/.git", package_path)).unwrap();
 }
 
-fn setup_move_home() -> (String, String) {
-    let home = home_dir().unwrap().to_string_lossy().to_string() + "/.move/test";
-    let credential_file = home.clone() + "/credential.toml";
-    (home, credential_file)
+fn setup_move_home(test_path: &str) -> (String, String) {
+    let mut move_home = home_dir().unwrap().to_string_lossy().to_string() + "/.move/test";
+    if !test_path.is_empty() {
+        move_home.push_str(&test_path);
+    }
+    let credential_file = move_home.clone() + "/credential.toml";
+    (move_home, credential_file)
 }
